requirements.txt
fastapi
uvicorn
requests
python-dotenv
google-genai
python-multipart
pandas